StyleTransfer/base_model_setup/변경 사항/requirements.txt
numpy==1.22.0
Pillow==10.2.0
protobuf==3.18.3
six==1.12.0
tensorboardX==1.8
torch==1.13.1
torchvision==0.14.1
tqdm==4.35.0
opencv-python==4.4.0.46
imageio==2.9.0